+++ src/inline_requests/generator.py
@@ -1,8 +1,14 @@
+import logging
+import warnings
+
 from functools import partial
 from types import GeneratorType
 
 from scrapy.http import Request
 from scrapy.utils.spider import iterate_spider_output
+
+
+logger = logging.getLogger(__name__)
 
 
 class RequestGenerator(object):
@@ -11,16 +17,32 @@ class RequestGenerator(object):
     """
 
     def __init__(self, callback, **kwargs):
+        """Initialize RequestGenerator.
+
+        Parameters
+        ----------
+        callback : callable
+            Callable callback (spider method).
+        **kwargs :
+            Extra callback keyword arguments.
+
+        """
         self.callback = callback
         self.kwargs = kwargs
 
     def __call__(self, response):
+        """Main response entry point.
+
+        This method calls the callback and wraps the returned generator.
+
+        """
         output = iterate_spider_output(self.callback(response=response, **self.kwargs))
         if not isinstance(output, GeneratorType):
             raise ValueError("Callback must return a generator type")
         return self._unwindGenerator(output)
 
     def _unwindGenerator(self, generator, _prev=None):
+        """Unwind (resume) generator."""
         while True:
             if _prev:
                 ret, _prev = _prev, None
@@ -29,11 +51,24 @@ class RequestGenerator(object):
                     ret = next(generator)
                 except StopIteration:
                     break
+
             if isinstance(ret, Request):
-                yield self._wrapRequest(ret, generator)
-                break
-            else:
-                yield ret
+                if ret.callback:
+                    warnings.warn("Got a request with callback set, bypassing "
+                                  "the generator wrapper. Generator may not "
+                                  "be able to resume. %s" % ret)
+                elif ret.errback:
+                    # By Scrapy defaults, a request without callback defaults to
+                    # self.parse spider method.
+                    warnings.warn("Got a request with errback set, bypassing "
+                                  "the generator wrapper. Generator may not "
+                                  "be able to resume. %s" % ret)
+                else:
+                    yield self._wrapRequest(ret, generator)
+                    return
+
+            # A request with callbacks, item or None object.
+            yield ret
 
     def _wrapRequest(self, request, generator):
         # Allowing existing callback or errbacks could lead to undesired